requirements.txt
numpy==1.20.3
kmodes==0.11.0
seaborn==0.11.1
matplotlib==3.3.4
pandas==1.3.3
PyYAML==5.4.1
scikit_learn==0.24.2
sweetviz==2.1.3
dill==0.3.4
ipython==7.27.0
nbconvert==6.1.0
pyyaml==5.4.1
openpyxl==3.0.7
dython==0.6.8
logging==0.5.1.2